Fixed quizzes allowing more required score to pass than is possible to achieve.

+++ frontend/toolkit-webapp/src/pages/ModuleCreator.js
@@ -124,15 +124,21 @@ function ModuleCreator(props) {
             if(gradeToPass<1){
                 setDialogText("The number of correct answers required must be greater than 0.")
                 handleOpenDialog()
-                //alert("File must be a PDF")
+                //alert("The number of correct answers required must be greater than 0.")
             } else{
 
                 console.log('works for Quiz')
                 var quiz = []
 
                 quiz = JSON.parse(sessionStorage.getItem("quiz"))
-                sessionStorage.clear()
-                onFinish({ title, type, description, quiz })
+                if(gradeToPass>quiz.length){
+                    setDialogText("The number of correct answers may not be greater than the number of questions.")
+                    handleOpenDialog()
+                    //alert("The number of correct answers may not be greater than the number of questions.")
+                } else{
+                    sessionStorage.clear()
+                    onFinish({ title, type, description, quiz })
+                }
             }
         } else if (type === 'PDF' && pdf !== null && typeof (pdf) !== 'undefined') {
             if (isPDF(pdf.name) === false) {
