+++ frontend/toolkit-webapp/src/pages/ModuleEditor.js
@@ -98,12 +98,25 @@ function ModuleEditor(props) {
         }
 
         if (type === 'Quiz' && JSON.parse(sessionStorage.getItem('quiz')) !== null) {
-            console.log('works for Quiz')
-            var quiz = []
+            if(gradeToPass<1){
+                setDialogText("The number of correct answers required must be greater than 0.")
+                handleOpenDialog()
+                //alert("The number of correct answers required must be greater than 0.")
+            } else{
 
-            quiz = JSON.parse(sessionStorage.getItem("quiz"))
-            sessionStorage.clear()
-            onFinish({ title, type, description, quiz, gradeToPass })
+                console.log('works for Quiz')
+                var quiz = []
+
+                quiz = JSON.parse(sessionStorage.getItem("quiz"))
+                if(gradeToPass>quiz.length){
+                    setDialogText("The number of correct answers may not be greater than the number of questions.")
+                    handleOpenDialog()
+                    //alert("The number of correct answers may not be greater than the number of questions.")
+                } else{
+                    sessionStorage.clear()
+                    onFinish({ title, type, description, quiz })
+                }
+            }
         }else if(type === 'PDF'){
             if(pdf !== null && typeof(pdf) !== 'undefined'){
                 if(isPDF(pdf.name) === false){
